// src/app/constants/states.enum.ts
export enum States {
  IDLE = 'IDLE',
  LOADING = 'LOADING',
  SUCCESS = 'SUCCESS',
  FAILURE = 'FAILURE',
  IS_AUTHENTICATED = 'IS_AUTHENTICATED',
  IS_ALREADY_AUTHENTICATED = 'IS_ALREADY_AUTHENTICATED',
  IS_NOT_AUTHENTICATED = 'IS_NOT_AUTHENTICATED',
  AUTHENTICATING = 'AUTHENTICATING',
  LOGGING_OUT = 'LOGGING_OUT'
}